fix: topic & comment controllers
