+++ packages/server/router/topic-routes.ts
@@ -3,6 +3,6 @@ import express from 'express'
 
 export const topicRoutes = express.Router()
 
-topicRoutes.get('/all', topicController.getTopics)
+topicRoutes.get('/', topicController.getTopics)
 topicRoutes.get('/:id', topicController.getTopicById)
-topicRoutes.post('/add', topicController.addTopic)
+topicRoutes.post('/', topicController.addTopic)
